% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Childrens Game FizzBuzz}
\usage{
FizzBuzz(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
a vector of numbers with corresponding numbers replaced by Fizz and Buzz.
}
\description{
This function takes a number and spits out a vector from 1 to that given number.
It replaces numbers divisible by 3 with "Fizz", numbers divisible by 5 with "Buzz",
and numbers divisible by both with "FizzBuzz".
}
\examples{
FizzBuzz(5)
FizzBuzz(20)
}
