% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather Data For Flagstaff, AZ}
\format{
a data frame with 25449 observations
\describe{
  \item{DATE}{The date of observation in the YYYY-MM-DD format.}
  \item{Precipitation}{The number of inches of precipitation on that day.}
  \item{SNOW}{The number of inches of snow on that day.}
  \item{TMAX}{The maximum temperature recorded that day, in degrees Fahrenheit.}
  \item{TMIN}{The minimun temperature recorded that day, in degrees Fahrenheit.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily temperatures, precipitation, and snowfall in Flagstaff, AZ, taken from the Pullman airport.
}
\keyword{datasets}
